add service logic

+++ shared/models/account.go
@@ -10,7 +10,7 @@ import (
 type Account struct {
 	ID        string    `json:"id"`
 	UserID    string    `json:"user_id"`
-	Balance   float64   `json:"balance"`
+	Balance   int64     `json:"balance"` // fixed scale (cents)
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
